recommender_system/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
fonttools==4.58.0
h11==0.16.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
lightfm==1.17
matplotlib==3.10.3
nest_asyncio==1.5.8
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==5.9.8
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.6.0
traitlets==5.14.3
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
